Load TOTP from rbw
This removes one dependency from `rofi-rbw`, which is nice. And the TOTP
is now calculated when you need it, not when the whole thing is loaded.
That means we can deal a lot better with the periodical changes.

Also, the code got a slight restructuring, fixing the interfaces better.

Issue #33

Co-Authored-By: Fabian Winter <5821180+fdw@users.noreply.github.com>
Co-Authored-By: tosti007 <brian-janssen@hotmail.com>