Fix frecency with `--no-folder`
With `--no-folder`, the returned `Entry` necessarily doesn't contain the
entry's folder. However, we do use it for the frecency hash. Thus, with
this option, we saved a hash without folder, but compared it with hashes
with the folder. That's not ideal 😉

The solution is to set the folder when requesting all data from `rbw`
and only saving to the cache afterwards.